+++ tests/testthat/test-formatter.r
@@ -1,5 +1,13 @@
 context("Formatters")
 
+test_that("time_format formats hms objects", {
+  a_time <- ISOdatetime(2012, 1, 1, 11, 30, 0, tz = "UTC")
+
+  expect_equal(time_format()(a_time), "11:30:00")
+  expect_equal(time_format()(hms::as.hms(a_time, tz = "UTC")), "11:30:00")
+  expect_equal(time_format(format = "%H")(hms::as.hms(a_time, tz = "UTC")), "11")
+})
+
 test_that("comma format always adds commas", {
   expect_equal(comma(1e3), "1,000")
   expect_equal(comma(1e6), "1,000,000")